add cart, fixing logic login. modeling for cart database

+++ lib/app/modules/home/model/product_model.dart
@@ -30,7 +30,7 @@ class Product {
       id: map['id'],
       code: map['code'],
       name: map['name'],
-      price: map['price'],
+      price: map['price'].toDouble(),
       stock: map['stock'],
     );
   }
